+++ apps/web/src/components/ui/chart.tsx
@@ -203,7 +203,7 @@ const ChartTooltipContent = React.forwardRef<
                                         )}
                                         <div
                                             className={cn(
-                                                "flex flex-1 justify-between leading-none",
+                                                "flex flex-1 justify-between leading-none gap-x-2",
                                                 nestLabel ? "items-end" : "items-center",
                                             )}
                                         >