+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stoker-platform/cli
 
+## 0.5.47
+
+### Patch Changes
+
+- feat: improve charts
+- Updated dependencies
+    - @stoker-platform/types@0.5.24
+    - @stoker-platform/node-client@0.5.36
+    - @stoker-platform/utils@0.5.30
+
 ## 0.5.46
 
 ### Patch Changes