+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.36
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.24
+    - @stoker-platform/utils@0.5.30
+
 ## 0.5.35
 
 ### Patch Changes