+++ packages/web-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/web-client
 
+## 0.5.32
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.24
+    - @stoker-platform/utils@0.5.30
+
 ## 0.5.31
 
 ### Patch Changes